+++ CHANGES.md
@@ -3,6 +3,8 @@
 ## 1.0.1 (...)
 - no longer explicitly enable the `ReferrerCheckPlugin` as it is enabled by
   default now and more intelligent
+- use external `fkooman/io` project for IO class, no longer embed our own
+  version
 
 ## 1.0.0 (2015-07-21)
 - initial release